+++ tap_shopify/streams/__init__.py
@@ -10,4 +10,5 @@ import tap_shopify.streams.custom_collections
 import tap_shopify.streams.locations
 import tap_shopify.streams.inventory_levels
 import tap_shopify.streams.inventory_items
+import tap_shopify.streams.product_variants
 import tap_shopify.streams.events