+++ tests/test_interrupted_sync.py
@@ -38,7 +38,6 @@ class InterruptedSyncTest(BaseTapTest):
 
         expected_streams = {'customers',
                             'events',
-                            'metafields',
                             'orders',
                             'products',
                             'transactions'}
@@ -89,12 +88,11 @@ class InterruptedSyncTest(BaseTapTest):
 
         base_state = {'bookmarks':
                      {'currently_sync_stream': currently_syncing_stream,
-                      'customers': {'updated_at': '2023-03-28T18:53:28.000000Z'},
-                      'events': {'created_at': '2023-01-22T05:05:53.000000Z'},
-                      'metafields': {'updated_at': '2023-01-07T21:18:05.000000Z'},
-                      'orders': {'updated_at': '2023-01-22T05:07:44.000000Z'},
-                      'products': {'updated_at': '2023-01-22T05:05:56.000000Z'},
-                      'transactions': {'created_at': '2022-06-26T00:06:38-04:00'}
+                      'customers': first_sync_state.get('bookmarks').get('customers'),
+                      'events': first_sync_state.get('bookmarks').get('events'),
+                      'orders': first_sync_state.get('bookmarks').get('orders'),
+                      'products': first_sync_state.get('bookmarks').get('products'),
+                      'transactions': first_sync_state.get('bookmarks').get('transactions')
                       }}
 
         # remove yet to be synced streams from base state and then set new state
@@ -196,18 +194,8 @@ class InterruptedSyncTest(BaseTapTest):
                 self.assertIsNotNone(resuming_bookmark_value)
                 self.assertTrue(self.is_expected_date_format(resuming_bookmark_value))
 
-                # verify the resuming bookmark is greater than 1st sync bookmark
-                # This is the expected behaviour for shopify as they are using date windowing
-                # TDL-17096 : Resuming bookmark value is getting assigned from execution time
-                # rather than the actual bookmark time for some streams.
-                # TODO transactions stream has equal bookmarks, orders stream has shown both equal
-                #   and greater than bookmark behavior, confirm if this is correct
-                if stream == 'transactions':
-                    self.assertEqual(resuming_bookmark_value, first_bookmark_value)
-                elif stream == 'orders':
-                    self.assertGreaterEqual(resuming_bookmark_value, first_bookmark_value)
-                else:
-                    self.assertGreater(resuming_bookmark_value, first_bookmark_value)
+                # verify the resuming bookmark is greater or equal than 1st sync bookmark
+                self.assertGreaterEqual(resuming_bookmark_value, first_bookmark_value)
 
                 # verify oldest record from resuming sync respects bookmark from previous sync
                 if stream in new_state['bookmarks'].keys() and resuming_sync_messages: